Added calls to display functions & calculation of avg slope
- each display function has a comment right next to it saying "call to display function" so that each display function can be found quickly
- for the avg slope calculation, how are we going to store the recent lux values?

+++ magnetometer_main/display.h
@@ -33,7 +33,7 @@ public:
     void show_file_name(char file_entry[], int index); // displays the screen to enter the file name
     void show_run_time(int run_time[], int index); // displays the screen to enter the run time wanted
     void show_test_ready(char file_name[], int run_time[]);
-    void show_test_in_progress(int run_time[], unsigned long time_elapsed, unsigned long recent_val, char file_name[], float avg_slope); // displays screen with time elapsed and active trendline
+    void show_test_in_progress(int run_time[], unsigned long time_elapsed, float recent_val, char file_name[], float avg_slope); // displays screen with time elapsed and active trendline
     void show_test_ended(char file_name[], unsigned long min, unsigned long sec); // displays test ended screen w/ file name and actual time elapsed
     void show_error_logger();
     void show_error_sensor();
